feat:create account

+++ src/services/api/index.js
@@ -1,8 +1,9 @@
 import axios from "axios";
-import {get} from "lodash";
+import {get, includes} from "lodash";
 import NProgress from "nprogress";
 import storage from "../storage";
 import config from "../../config";
+import Swal from "sweetalert2";
 
 
 NProgress.configure({
@@ -27,16 +28,37 @@ request.interceptors.request.use((config) => {
     return config;
 }, (error) => {
     NProgress.done(true);
-    console.log(error)
+    return Promise.reject(error);
 });
 
 request.interceptors.response.use((response) => {
     NProgress.done(true);
     return response;
 }, (error) => {
-    const statusCode = error.response.status;
+    const statusCode = error?.response?.status;
     if (statusCode == 401) {
-        window.localStorage.clear();
+        if(!includes(window.location.pathname,'auth')) {
+            Swal.fire({
+                title: 'Your token expired',
+                icon: 'error',
+                backdrop: 'rgba(0,0,0,0.9)',
+                background: 'none',
+                confirmButtonColor: '#CD4472',
+                confirmButtonText: 'Logout',
+                customClass: {
+                    title: 'title-color',
+                    content: 'text-color',
+                    icon: 'icon-color',
+                },
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    window.localStorage.clear();
+                    window.location.href='/auth';
+                }
+            });
+        }else{
+            window.localStorage.clear();
+        }
     }
 
     NProgress.done(true);
